refactor: split cache layers from DB modules, inline voiceAdapter, add qlty hook
- Extract src/db/customCommandCache.ts (~255 lines) from customCommands.ts,
  separating the cache infrastructure and lookup functions from CRUD queries.
  customCommands.ts shrinks from 510 to ~160 lines.

- Extract src/db/counterCache.ts from counters.ts, separating the counter
  lookup cache and isCounterCommandTaken from CRUD/increment/archive functions.
  counters.ts shrinks from 346 to ~200 lines.

- Inline src/voiceAdapter.ts into audioPlayer.ts: buildAdapter() had a single
  consumer; the adapter is now a private implementation detail of connect().

- Add .claude/hooks/pre-commit-quality-check.sh and update settings.json to
  run qlty on staged files before every git commit (ported from main).

All public re-exports in src/db.ts are unchanged so callers are unaffected.
