+++ src/audioPlayer.ts
@@ -9,13 +9,14 @@ import {
   type AudioResource,
   type VoiceConnection,
   type AudioPlayer as DjsAudioPlayer,
+  type DiscordGatewayAdapterCreator,
+  type DiscordGatewayAdapterLibraryMethods,
 } from '@discordjs/voice';
-import { Client, ChannelType } from 'discord.js';
+import { Client, ChannelType, type VoiceBasedChannel } from 'discord.js';
 import { DISCORD_GUILD_ID, DISCORD_VOICE_CHANNEL_ID } from './config';
 import { setVoiceConnected, setVoiceDisconnected, setVoiceIdle } from './statusStore';
 
 const log = createLogger('AudioPlayer');
-import { buildAdapter } from './voiceAdapter';
 import { isPermanentVoiceMisconfigurationError } from './discordUtils';
 import {
   type ConnectionHandlerDeps,
@@ -34,6 +35,60 @@ let shouldAutoReconnect = false;
 let currentAttemptId = 0;
 let currentChannelId: string | null = null;
 let targetChannelId: string | undefined = undefined;
+
+// ─── Voice adapter ────────────────────────────────────────────────────────────
+// Bypasses discord.js's built-in voiceAdapterCreator to avoid type/version
+// incompatibilities with discord.js v14. Listens to raw gateway events instead.
+
+// Tracks the cleanup function for the currently active adapter so it can be
+// removed before a new one is registered, preventing listener accumulation.
+let activeAdapterCleanup: (() => void) | null = null;
+
+function buildAdapter(channel: VoiceBasedChannel): DiscordGatewayAdapterCreator {
+  return (methods: DiscordGatewayAdapterLibraryMethods) => {
+    if (activeAdapterCleanup) {
+      activeAdapterCleanup();
+      activeAdapterCleanup = null;
+    }
+
+    function onRaw(packet: { t: string; d: Record<string, unknown> }) {
+      if (packet.t === 'VOICE_STATE_UPDATE') {
+        methods.onVoiceStateUpdate(packet.d as unknown as Parameters<typeof methods.onVoiceStateUpdate>[0]);
+      }
+      if (packet.t === 'VOICE_SERVER_UPDATE') {
+        methods.onVoiceServerUpdate(packet.d as unknown as Parameters<typeof methods.onVoiceServerUpdate>[0]);
+      }
+    }
+
+    channel.client.setMaxListeners(channel.client.getMaxListeners() + 1);
+    channel.client.on('raw', onRaw);
+
+    function cleanup(): void {
+      channel.client.off('raw', onRaw);
+      channel.client.setMaxListeners(Math.max(channel.client.getMaxListeners() - 1, 0));
+      if (activeAdapterCleanup === cleanup) {
+        activeAdapterCleanup = null;
+      }
+    }
+
+    activeAdapterCleanup = cleanup;
+
+    return {
+      sendPayload: (payload: unknown) => {
+        try {
+          // eslint-disable-next-line @typescript-eslint/no-explicit-any
+          channel.guild.shard.send(payload as any);
+          return true;
+        } catch {
+          return false;
+        }
+      },
+      destroy: cleanup,
+    };
+  };
+}
+
+// ─── Reconnect ────────────────────────────────────────────────────────────────
 
 const RECONNECT_BASE_DELAY_MS = 5_000;
 const RECONNECT_MAX_DELAY_MS = 60_000;
